refactor(vault): simplify the geo-blocking
