--- services/vault/src/api/error/index.ts
@@ -1,9 +1,0 @@
-export const isError451 = (error: unknown): boolean => {
-  if (!error || typeof error !== "object") {
-    return false;
-  }
-  const maybeWithStatus = error as { status?: unknown };
-  return (
-    typeof maybeWithStatus.status === "number" && maybeWithStatus.status === 451
-  );
-};
