--- services/vault/src/api/index.ts
@@ -1,4 +1,0 @@
-export { isError451 } from "./error";
-export { fetchHealthCheck } from "./healthCheckClient";
-export type { HealthCheckResponse } from "./healthCheckClient";
-export { ApiError } from "./types";
