--- services/vault/src/components/HealthCheckWrapper.tsx
@@ -1,9 +1,0 @@
-import type { PropsWithChildren } from "react";
-
-import { useHealthCheck } from "@/hooks/useHealthCheck";
-
-export function HealthCheckWrapper({ children }: PropsWithChildren) {
-  useHealthCheck();
-
-  return <>{children}</>;
-}
