+++ services/vault/src/services/health/index.ts
@@ -3,6 +3,8 @@ export {
   checkGraphQLEndpoint,
   createEnvConfigError,
   createWagmiInitError,
+  fetchHealthCheck,
   runHealthChecks,
+  type HealthCheckResponse,
   type HealthCheckResult,
 } from "./healthCheckService";
